Move more utils
