+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/features.js
@@ -1,98 +1,105 @@
-/*
- * When we insert in the DB we need to translate the property names that either come from a CTN layer
- * or the draw tools into columns names in the database
+import { useMemo } from "react";
+import { v4 as uuidv4 } from "uuid";
+import { getIntersectionLabel } from "./utils";
+
+export const useComponentFeatureCollection = (component) =>
+  useMemo(() => {
+    if (!component || !component?.features) return null;
+    return { type: "FeatureCollection", features: component.features };
+  }, [component]);
+
+/**
+ * Captured from feature object example
+ * {
+ *    "geometry": {
+ *        "type": "LineString",
+ *        "coordinates": [],
+ *    },
+ *    "properties": {
+ *        "OBJECTID": 20041,
+ *        "CTN_SEGMENT_ID": 153083,
+ *        "FULL_STREET_NAME": "W MARTIN LUTHER KING JR BLVD",
+ *        "FROM_ADDRESS_MIN": 300,
+ *        "TO_ADDRESS_MAX": 303,
+ *        "LINE_TYPE": "On-Street",
+ *        "GLOBALID": "cb89d800-d6a3-491f-91e6-61cab17ac738",
+ *        "Shape__Length": 60.3716805891339,
+ *        "id": 20041,
+ *        "_layerId": "ctn-lines-underlay",
+ *        "_label": "300 BLK W MARTIN LUTHER KING JR BLVD"
+ *    }
+ * }
  */
-export const featureTableFieldMap = {
-  feature_street_segments: {
-    full_street_name: "FULL_STREET_NAME",
-    ctn_segment_id: "CTN_SEGMENT_ID",
-    from_address_min: "FROM_ADDRESS_MIN",
-    to_address_max: "TO_ADDRESS_MAX",
-    line_type: "LINE_TYPE",
-    // project_extent_id: undefined,
-    // knack_id: undefined,
-    // symbol: undefined,
-    // render_type: undefined,
-    // name: undefined
-  },
-  feature_intersections: {
-    intersection_id: "COA_INTERSECTION_POINTS_ID",
-    // project_extent_id: undefined,
-    // render_type: undefined,
-    // name: undefined
-  },
-  feature_signals: {
-    // location_name: undefined,
-    // knack_id: undefined,
-    // project_extent_id: undefined,
-    // render_type: undefined,
-    // name: undefined
-    // signal_type: undefined
-  },
-  feature_drawn_lines: {},
-  feature_drawn_points: {},
-};
 
-export const makeLineStringFeatureInsertionData = (
-  featureTable,
-  draftComponent,
-  featuresToInsert
+/**
+ * Make a feature object from a clicked CTN feature and AGOL returned data about it
+ * @param {Object} featureFromAgolGeojson - a feature object from AGOL
+ * @param {Object} clickedFeature - a feature object from a Mapbox click event
+ * @param {Object} ctnLinesGeojson - a feature collection of CTN lines currently in map view
+ * @returns {Object} - a feature object
+ */
+export const makeCapturedFromLayerFeature = (
+  featureFromAgolGeojson,
+  clickedFeature,
+  ctnLinesGeojson
 ) => {
-  draftComponent.features.forEach((feature) => {
-    const featureToInsert = {};
-    const translationMap = featureTableFieldMap[featureTable];
+  const newFeature = {
+    geometry: featureFromAgolGeojson.geometry,
+    properties: {
+      ...featureFromAgolGeojson.properties,
+      id: featureFromAgolGeojson.id,
+      // AGOL data doesn't include source layer ID so we grab it from the clicked Mapbox feature
+      _layerId: clickedFeature.layer.source,
+    },
+  };
 
-    Object.keys(translationMap).forEach((key) => {
-      const translatedKey = translationMap[key];
+  if (newFeature.properties._layerId.includes("point")) {
+    newFeature.properties._label = getIntersectionLabel(
+      newFeature,
+      ctnLinesGeojson
+    );
+  } else {
+    newFeature.properties._label = `${newFeature.properties.FROM_ADDRESS_MIN} BLK ${newFeature.properties.FULL_STREET_NAME}`;
+  }
 
-      featureToInsert[key] = feature.properties[translatedKey];
-    });
-
-    // Add source_layer
-    featureToInsert["source_layer"] = feature.properties._layerId;
-
-    // Convert from LineString to a MultiLineString
-    const coordinatesArray = feature.geometry.coordinates;
-
-    featureToInsert["geography"] = {
-      type: "MultiLineString",
-      coordinates: [coordinatesArray],
-    };
-
-    featuresToInsert.push(featureToInsert);
-  });
-
-  return featuresToInsert;
+  return newFeature;
 };
 
-export const makePointFeatureInsertionData = (
-  featureTable,
-  draftComponent,
-  featuresToInsert
-) => {
-  draftComponent.features.forEach((feature) => {
-    const featureToInsert = {};
-    const translationMap = featureTableFieldMap[featureTable];
+/**
+ * Drawn feature object example
+ * {
+ *   "id": "9e0c688733283be4991cb9df39ae284d",
+ *   "geometry": { coordinates: [], type: "LineString" },
+ *   "type": "Feature",
+ *   "properties": {
+ *     "DRAW_ID": "88230580-099f-4c93-b8e5-ad3086caf0e6",
+ *     "_layerId": "drawnByUserLine",
+ * }
+ */
 
-    Object.keys(translationMap).forEach((key) => {
-      const translatedKey = translationMap[key];
+/**
+ * Make a drawn feature object from GeoJSON output of Mapbox GL Draw
+ * @param {Object} feature - GeoJSON feature object from Mapbox GL Draw onCreate
+ * @param {String} linkMode - tracks if we are editing "lines" or "points"
+ * @returns {Object} - a feature object
+ */
+export const makeDrawnFeature = (feature, linkMode) => {
+  feature.properties["DRAW_ID"] = uuidv4();
 
-      featureToInsert[key] = feature.properties[translatedKey];
-    });
-
-    // Add source_layer
-    featureToInsert["source_layer"] = feature.properties._layerId;
-
-    // Convert from Point to a MultiPoint
-    const coordinatesArray = feature.geometry.coordinates;
-
-    featureToInsert["geography"] = {
-      type: "MultiPoint",
-      coordinates: [coordinatesArray],
-    };
-
-    featuresToInsert.push(featureToInsert);
-  });
-
-  return featuresToInsert;
+  // TODO: Current schema uses "drawnByUser" for lines but we need _layerId to contain
+  // "line" or "point" to the popup to work as expected (otherwise mutliple components
+  // show up in the popup when clicking a feature)
+  if (linkMode === "lines") {
+    feature.properties["_layerId"] = "drawnByUserLine";
+  } else if (linkMode === "points") {
+    feature.properties["_layerId"] = "drawnByUserPoint";
+  }
 };
+
+/**
+ * Determine if a feature is drawn by the user
+ * @param {Object} feature
+ * @returns {Boolean} - true if feature is drawn by user
+ */
+export const isDrawnFeature = (feature) =>
+  feature?.properties?.DRAW_ID ? true : false;
