+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/makeFeatures.js
@@ -1,0 +1,98 @@
+/*
+ * When we insert in the DB we need to translate the property names that either come from a CTN layer
+ * or the draw tools into columns names in the database
+ */
+export const featureTableFieldMap = {
+  feature_street_segments: {
+    full_street_name: "FULL_STREET_NAME",
+    ctn_segment_id: "CTN_SEGMENT_ID",
+    from_address_min: "FROM_ADDRESS_MIN",
+    to_address_max: "TO_ADDRESS_MAX",
+    line_type: "LINE_TYPE",
+    // project_extent_id: undefined,
+    // knack_id: undefined,
+    // symbol: undefined,
+    // render_type: undefined,
+    // name: undefined
+  },
+  feature_intersections: {
+    intersection_id: "COA_INTERSECTION_POINTS_ID",
+    // project_extent_id: undefined,
+    // render_type: undefined,
+    // name: undefined
+  },
+  feature_signals: {
+    // location_name: undefined,
+    // knack_id: undefined,
+    // project_extent_id: undefined,
+    // render_type: undefined,
+    // name: undefined
+    // signal_type: undefined
+  },
+  feature_drawn_lines: {},
+  feature_drawn_points: {},
+};
+
+export const makeLineStringFeatureInsertionData = (
+  featureTable,
+  draftComponent,
+  featuresToInsert
+) => {
+  draftComponent.features.forEach((feature) => {
+    const featureToInsert = {};
+    const translationMap = featureTableFieldMap[featureTable];
+
+    Object.keys(translationMap).forEach((key) => {
+      const translatedKey = translationMap[key];
+
+      featureToInsert[key] = feature.properties[translatedKey];
+    });
+
+    // Add source_layer
+    featureToInsert["source_layer"] = feature.properties._layerId;
+
+    // Convert from LineString to a MultiLineString
+    const coordinatesArray = feature.geometry.coordinates;
+
+    featureToInsert["geography"] = {
+      type: "MultiLineString",
+      coordinates: [coordinatesArray],
+    };
+
+    featuresToInsert.push(featureToInsert);
+  });
+
+  return featuresToInsert;
+};
+
+export const makePointFeatureInsertionData = (
+  featureTable,
+  draftComponent,
+  featuresToInsert
+) => {
+  draftComponent.features.forEach((feature) => {
+    const featureToInsert = {};
+    const translationMap = featureTableFieldMap[featureTable];
+
+    Object.keys(translationMap).forEach((key) => {
+      const translatedKey = translationMap[key];
+
+      featureToInsert[key] = feature.properties[translatedKey];
+    });
+
+    // Add source_layer
+    featureToInsert["source_layer"] = feature.properties._layerId;
+
+    // Convert from Point to a MultiPoint
+    const coordinatesArray = feature.geometry.coordinates;
+
+    featureToInsert["geography"] = {
+      type: "MultiPoint",
+      coordinates: [coordinatesArray],
+    };
+
+    featuresToInsert.push(featureToInsert);
+  });
+
+  return featuresToInsert;
+};
